BaseCardView extracted out and DashboardCardView separated as well
